# lib/api-client-react/src/generated/api.schemas.ts/createRouterBody.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * VoucherNet API
 * OpenAPI spec version: 1.0.0
 */

export interface CreateRouterBody {
  name: string;
  host: string;
  port?: number;
  username: string;
  password: string;
  isActive?: boolean;
}