# lib/api-client-react/src/generated/api.schemas.ts/dashboardStats.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * VoucherNet API
 * OpenAPI spec version: 1.0.0
 */
import type { Voucher } from './voucher';

export interface DashboardStats {
  totalVouchers: number;
  unprintedVouchers: number;
  printedVouchers: number;
  routerCount: number;
  activeRouters: number;
  recentVouchers?: Voucher[];
}