# lib/api-client-react/src/generated/api.schemas.ts/generateVouchersBody.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * VoucherNet API
 * OpenAPI spec version: 1.0.0
 */
import type { GenerateVouchersBodyPrefix } from './generateVouchersBodyPrefix';
import type { GenerateVouchersBodyComment } from './generateVouchersBodyComment';
import type { GenerateVouchersBodyServer } from './generateVouchersBodyServer';

export interface GenerateVouchersBody {
  routerId: number;
  profile: string;
  /**
   * @minimum 1
   * @maximum 200
   */
  qty: number;
  prefix?: GenerateVouchersBodyPrefix;
  comment?: GenerateVouchersBodyComment;
  server?: GenerateVouchersBodyServer;
}