# lib/api-client-react/src/generated/api.schemas.ts/hotspotProfile.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * VoucherNet API
 * OpenAPI spec version: 1.0.0
 */
import type { HotspotProfileRateLimit } from './hotspotProfileRateLimit';
import type { HotspotProfileValidity } from './hotspotProfileValidity';
import type { HotspotProfilePrice } from './hotspotProfilePrice';
import type { HotspotProfileSharedUsers } from './hotspotProfileSharedUsers';
import type { HotspotProfileAddrPool } from './hotspotProfileAddrPool';

export interface HotspotProfile {
  name: string;
  rateLimit?: HotspotProfileRateLimit;
  validity?: HotspotProfileValidity;
  price?: HotspotProfilePrice;
  sharedUsers?: HotspotProfileSharedUsers;
  addrPool?: HotspotProfileAddrPool;
  lockMac?: boolean;
}