# lib/api-client-react/src/generated/api.schemas.ts/hotspotSession.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * VoucherNet API
 * OpenAPI spec version: 1.0.0
 */
import type { HotspotSessionMacAddress } from './hotspotSessionMacAddress';
import type { HotspotSessionBytesIn } from './hotspotSessionBytesIn';
import type { HotspotSessionBytesOut } from './hotspotSessionBytesOut';
import type { HotspotSessionServer } from './hotspotSessionServer';

export interface HotspotSession {
  user: string;
  address: string;
  macAddress?: HotspotSessionMacAddress;
  uptime: string;
  bytesIn?: HotspotSessionBytesIn;
  bytesOut?: HotspotSessionBytesOut;
  server?: HotspotSessionServer;
}