# lib/api-client-react/src/generated/api.schemas.ts/testResult.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * VoucherNet API
 * OpenAPI spec version: 1.0.0
 */
import type { TestResultRouterBoard } from './testResultRouterBoard';
import type { TestResultVersion } from './testResultVersion';

export interface TestResult {
  success: boolean;
  message: string;
  routerBoard?: TestResultRouterBoard;
  version?: TestResultVersion;
}