# lib/api-client-react/src/generated/api.schemas.ts/voucher.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * VoucherNet API
 * OpenAPI spec version: 1.0.0
 */
import type { VoucherComment } from './voucherComment';
import type { VoucherPrintedAt } from './voucherPrintedAt';

export interface Voucher {
  id: number;
  routerId: number;
  username: string;
  password: string;
  profileName: string;
  price: string;
  validity: string;
  comment?: VoucherComment;
  printedAt?: VoucherPrintedAt;
  createdAt: string;
}