# lib/api-client-react/src/generated/api.schemas.ts/voucherListResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * VoucherNet API
 * OpenAPI spec version: 1.0.0
 */
import type { Voucher } from './voucher';

export interface VoucherListResponse {
  vouchers: Voucher[];
  total: number;
}